+++ src/Terrabuild.Tests/SourceControls/Repository.fs
@@ -1,0 +1,11 @@
+module Terrabuild.Tests.RepositorySourceControls
+
+open FsUnit
+open NUnit.Framework
+
+[<TestCase("git@github.com:MagnusOpera/terrabuild.git", "MagnusOpera/terrabuild")>]
+[<TestCase("https://github.com/MagnusOpera/terrabuild.git", "MagnusOpera/terrabuild")>]
+[<TestCase("ssh://git@github.com/MagnusOpera/terrabuild.git", "MagnusOpera/terrabuild")>]
+[<TestCase("https://gitlab.example.com/group/subgroup/repo.git", "gitlab.example.com/group/subgroup/repo")>]
+let ``normalize repository identity`` repository expected =
+    Git.tryNormalizeRepositoryIdentity repository |> should equal (Some expected)